requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.13
django-braces==1.13.0
django-ckeditor==5.9.0
django-extensions==3.0.5
django-nested-admin==3.3.2
django-prettyjson==0.4.1
django-tables2==2.3.1
gunicorn==20.0.4
idna==2.10
lxml==4.5.2
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
standardjson==0.3.1
urllib3==1.25.10
whitenoise==5.1.0

# Additional requirements
# Required for LTI authentication
git+https://github.com/Harvard-University-iCommons/django-auth-lti.git@v2.0.2#egg=django-auth-lti
lti>=0.9.4